Fixed warnings

+++ i2c.c
@@ -17,7 +17,7 @@ int main() {
   // Range is 0x03 to 0x77
   // 0x6A is the slave address for the accelerometer and gyro on an LSM9DS1
   // with the SA0 line pulled low
-  int addr = 0x6A;
+  uint8_t addr = 0x6A;
   set_slave(i2c,addr);
 
   if(i2c == -1) // Something terrible has happened
